Extract OpenTelemetry test
