+++ tests/Common/TracerTestTrait.php
@@ -463,8 +463,14 @@ trait TracerTestTrait
     /**
      * Returns the raw response body, if any, or null otherwise.
      */
-    public function retrieveDumpedData()
+    public function retrieveDumpedData(callable $until = null, $throw = false)
     {
+        if (!$until) {
+            $until = function ($request) {
+                return (strpos($request["uri"] ?? "", "/telemetry/") !== 0);
+            };
+        }
+
         $allResponses = [];
 
         // When tests run with the background sender enabled, there might be some delay between when a trace is flushed
@@ -488,12 +494,18 @@ trait TracerTestTrait
                 $loaded = json_decode($response, true);
                 array_push($allResponses, ...$loaded);
                 foreach ($loaded as $request) {
-                    if (strpos($request["uri"] ?? "", "/telemetry/") !== 0) {
-                        break 2;
+                    if ($until($request)) {
+                        return $allResponses;
                     }
                 }
+                \usleep(1000);
             }
         }
+
+        if ($throw) {
+            throw new \LogicException('The expected request was not found');
+        }
+
         return $allResponses;
     }
