+++ tests/Frameworks/Custom/OpenTelemetry/composer.json
@@ -1,0 +1,6 @@
+{
+    "require": {
+        "php": ">=7.4",
+        "open-telemetry/sdk": "^1.0"
+    }
+}
