+++ tests/Frameworks/Custom/OpenTelemetry/index.php
@@ -1,0 +1,16 @@
+<?php
+
+use OpenTelemetry\SDK\Trace\TracerProvider;
+
+require __DIR__ . '/vendor/autoload.php';
+
+try {
+    $tracerProvider = new TracerProvider();
+    $tracer = $tracerProvider->getTracer('foo');
+    $span = $tracer->spanBuilder('barbar')
+        ->startSpan()
+    ;
+    $span->end();
+} finally {
+    \dd_trace_internal_fn("finalize_telemetry");
+}
