+++ tests/Frameworks/Custom/Version_Autoloaded/composer.json
@@ -1,18 +1,11 @@
 {
     "require": {
         "php": ">=5.4",
-        "nikic/fast-route": "^1.3",
-        "open-telemetry/sdk": "^1.0",
-        "opentracing/opentracing": "^1.0"
+        "nikic/fast-route": "^1.3"
     },
     "autoload": {
         "psr-4": {
             "App\\": "src/"
-        }
-    },
-    "config": {
-        "allow-plugins": {
-            "php-http/discovery": false
         }
     }
 }
