+++ tests/phpunit.xml
@@ -60,6 +60,7 @@
             <directory>./OpenTelemetry/Integration/Logs</directory>
             <directory>./OpenTelemetry/Integration/SDK</directory>
             <file>./OpenTelemetry/Integration/InteroperabilityTest.php</file>
+            <file>./OpenTelemetry/Integration/InternalTelemetryTest.php</file>
         </testsuite>
         <testsuite name="slim-312-test">
             <directory>./Integrations/Slim/V3_12</directory>
